feat: UMP-only API with CoreMIDI MIDI 2.0 EventList I/O
Breaking: send/receive native-endian u32 UMP words instead of MIDI 1.0 bytes.
Apple backends use Protocol::Midi20; other OS backends stub connect/send until
docs/ump-backend-compliance.md is green. No midi2 crate dependency.

+++ src/os/unix.rs
@@ -18,7 +18,7 @@ where
         data: T,
     ) -> Result<MidiInputConnection<T>, ConnectError<Self>>
     where
-        F: FnMut(u64, &[u8], &mut T) + Send + 'static;
+        F: FnMut(u64, &[u32], &mut T) + Send + 'static;
 }
 
 /// Trait that is implemented by `MidiOutput` on platforms that
